Create README.md
Hi I'm Dev Sisodiya a passionate programmer driven by innovation. Frustrated with the limitations of apps and websites, I embarked on a mission to revolutionize the tech industry. With a keen eye for improvement, I'm dedicated to crafting cutting-edge solutions that bridge the gap between functionality and user experience. Join me on my journey to shape the digital landscape!!